refactor: moves util functions into utils directory
