+++ packages/osc-academic-hub/app/routes/login.tsx
@@ -5,7 +5,8 @@ import * as React from 'react';
 
 import { createUserSession, getUserId } from '~/session.server';
 import { verifyLogin } from '~/models/user.server';
-import { safeRedirect, validateEmail } from '~/utils';
+import { validateEmail } from '~/utils/_tmp_/validateEmail';
+import { safeRedirect } from '~/utils/_tmp_/safeRedirect';
 
 export const loader: LoaderFunction = async ({ request }) => {
     const userId = await getUserId(request);
