+++ packages/osc-ecommerce/app/routes/join.tsx
@@ -6,7 +6,8 @@ import * as React from 'react';
 import { getUserId, createUserSession } from '~/session.server';
 
 import { createUser, getUserByEmail } from '~/models/user.server';
-import { safeRedirect, validateEmail } from '~/utils';
+import { safeRedirect } from '~/utils/_tmp_/safeRedirect';
+import { validateEmail } from '~/utils/_tmp_/validateEmail';
 
 export const loader: LoaderFunction = async ({ request }) => {
     const userId = await getUserId(request);
