+++ packages/osc-ecommerce/app/utils/_tmp_/user.test.ts
@@ -1,4 +1,4 @@
-import { validateEmail } from './utils';
+import { validateEmail } from './validateEmail';
 
 test('validateEmail returns false for non-emails', () => {
     expect(validateEmail(undefined)).toBe(false);
