+++ packages/osc-ecommerce/app/utils/_tmp_/validateEmail.ts
@@ -1,0 +1,3 @@
+export function validateEmail(email: unknown): email is string {
+    return typeof email === 'string' && email.length > 3 && email.includes('@');
+}
